update pyproject.toml, move function
