+++ pyproject.toml
@@ -26,6 +26,7 @@ dependencies = [
   "seaborn",
   "scikit-learn",
   "statsmodels",
+  "matplotlib",
   "yattag"
 ]
